Support explicit module versions
